Fix a build error
This tries to revert https://github.com/mozilla/OneCRL-Tools/pull/105

I'm not sure why we changed from `go get` to `go install` in that PR,
but it doesn't work when we try to build the lambda. See the error in
https://bugzilla.mozilla.org/show_bug.cgi?id=1397882#c4

IIRC, neither `go install` nor `go build` will download the dependencies,
only `go get` does that. If we were trying to avoid always building the
"master", then we can go get a relative path.